web_api_app::routes::bucket
Module bucket_function

This mod has s3-related functions.

Functions

generate_pre_signed_url_for_upload

get_days

Read the days that existing items are narrowed down by year and month in the s3 bucket.

get_months

Read the month that existing items are narrowed down by year in the s3 bucket.

get_objects

Read the objects that existing items are narrowed down by year, month and day in the s3 bucket.

get_years

Read the years that exist items in the s3 bucket.